requirements.txt
attrs==23.2.0
gspread-pandas==3.3.0
google-auth==2.32.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
numpy==2.0.1
pandas==2.2.2
python-chess==1.999
requests==2.32.3
